requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
supabase==2.9.1
python-dotenv==1.0.1
httpx>=0.27.2
pydantic==2.10.0
python-multipart==0.0.12
scholarly==1.7.11
requests